created package conf and pagestructs. Conf holds things to do with configuration and pagestructs holds all structs for pages and forms
