+++ cmd/main.go
@@ -5,15 +5,16 @@ import (
 	"fmt"
 	"html/template"
 	"io"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"strconv"
 
 	"github.com/gorilla/sessions"
 	_ "github.com/mattn/go-sqlite3"
+	"github.com/vtallen/go-link-shortener/internal/conf"
+	"github.com/vtallen/go-link-shortener/internal/pagestructs"
+	"github.com/vtallen/go-link-shortener/internal/sessmngt"
 	"github.com/vtallen/go-link-shortener/pkg/codegen"
-	"gopkg.in/yaml.v2"
 
 	"github.com/labstack/echo-contrib/session"
 	"github.com/labstack/echo/v4"
@@ -41,89 +42,15 @@ func newTemplate() *Templates {
  */
 type IndexData struct {
 	ShortcodeForm ShortcodeForm
-	Server        *Server
-}
-
-/*
-* Name: LoginData
-*
-* Description: This struct is used to pass data to the login page.
- */
-type LoginData struct {
-	LoginForm       LoginForm
-	HasError        bool
-	ErrorText       string
-	AlreadyLoggedIn bool
-}
-
-type LoginForm struct {
-	Email    string
-	Password string
-}
-
-type RegisterData struct {
-	RegisterForm RegisterForm
-	HasError     bool
-	ErrorText    string
-	Success      bool
-}
-
-type RegisterForm struct {
-	Email    string
-	Password string
+	Server        *conf.Server
 }
 
 type UserPageData struct{}
-
-type ErrorPageData struct {
-	ErrorText string
-}
 
 type ShortcodeForm struct {
 	URL      string
 	Result   string
 	HasError bool
-}
-type Shortcodes struct {
-	ShortcodeLength int    `yaml:"shortcode_length"`
-	Universe        string `yaml:"shortcode_universe"`
-}
-
-type Auth struct {
-	ApiKeyLen        int    `yaml:"api_key_length"`
-	RootUsername     string `yaml:"root_username"`
-	RootPassword     string `yaml:"root_password"`
-	TLSCert          string `yaml:"tls_cert"`
-	TLSKey           string `yaml:"tls_key"`
-	CookieMaxAgeDays int    `yaml:"cookie_max_age_days"`
-	CookieSecret     string `yaml:"cookie_secret"`
-}
-
-type Server struct {
-	Host string `yaml:"host"`
-	Port int    `yaml:"port"`
-}
-
-type Config struct {
-	Shortcodes Shortcodes
-	Auth       Auth
-	Server     Server
-}
-
-func LoadConfig(filename string) (*Config, error) {
-	data, err := ioutil.ReadFile(filename)
-	if err != nil {
-		return nil, err
-	}
-
-	var config Config
-	if err := yaml.Unmarshal(data, &config); err != nil {
-		return nil, err
-	}
-
-	fmt.Println(config)
-
-	return &config, nil
 }
 
 func main() {
@@ -139,7 +66,7 @@ func main() {
 	e := echo.New() // Create the web server
 
 	// config := NewConfig()
-	config, err := LoadConfig("config.yaml")
+	config, err := conf.LoadConfig("config.yaml")
 	if err != nil {
 		e.Logger.Fatal(err.Error())
 	}
@@ -173,7 +100,7 @@ func main() {
 	// Setup data structs for the different pages
 	indexData := IndexData{}
 	indexData.Server = &config.Server
-	errorPageData := ErrorPageData{"No error"}
+	errorPageData := pagestructs.ErrorPageData{"No error"}
 
 	// Serve the index page
 	e.GET("/", func(c echo.Context) error {
@@ -192,31 +119,31 @@ func main() {
 		return HandleAddLink(c, db, config, &indexData)
 	})
 
-	loginData := LoginData{} // Data used by login/register pages
+	loginData := pagestructs.LoginData{} // Data used by login/register pages
 
 	e.GET("/login", func(c echo.Context) error {
 		loginData.HasError = false
 		loginData.ErrorText = ""
 		loginData.LoginForm.Email = ""
 
-		return HandleLoginPage(c, &loginData, config)
+		return sessmngt.HandleLoginPage(c, &loginData, config)
 	})
 	e.POST("/login", func(c echo.Context) error {
-		return HandleLoginSession(c, db, &loginData, config)
+		return sessmngt.HandleLoginSession(c, db, &loginData, config)
 	})
 
 	e.GET("/logout", func(c echo.Context) error {
-		return HandleLogout(c, config)
+		return sessmngt.HandleLogout(c, config)
 	})
 
-	registerData := RegisterData{}
+	registerData := pagestructs.RegisterData{}
 	e.GET("/register", func(c echo.Context) error {
 		loginData.HasError = false
 		loginData.ErrorText = ""
-		return HandleRegisterPage(c, &registerData, config)
+		return sessmngt.HandleRegisterPage(c, &registerData, config)
 	})
 	e.POST("/register", func(c echo.Context) error {
-		return HandleRegisterSession(c, db, &registerData, config)
+		return sessmngt.HandleRegisterSession(c, db, &registerData, config)
 	})
 
 	userPageData := UserPageData{}
